+++ sdk/notificationhubs/notification-hubs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Added support for [Live Activity Apple message types](https://developer.apple.com/documentation/activitykit/updating-and-ending-your-live-activity-with-activitykit-push-notifications).
+- Added typings for APNs and WNS headers via the `createAppleNotification` and the `createWindowsNotification` factory methods.
+
 ### Breaking Changes
 
 ### Bugs Fixed